+++ CHANGELOG.md
@@ -1,8 +1,20 @@
-NGELOG
+# CHANGELOG
 
 The changelog for `Accelerator-TextChat` iOS.
 
 --------------------------------------
+
+1.0.3
+-----
+
+### Fixes
+
+- Fix a bug that `nil` message is treated as a successful case.
+- Fix a file reference issue on Subspec
+
+### Breaking changes 
+
+- Remove `receiverAlias` property on `OTTextChat` as the information is available on `OTTextMessage`
 
 1.0.2
 -----